fixed linting; organized files
