+++ dashboard-ui/src/components/SurveyResults/surveyResults.css
@@ -145,151 +145,3 @@
     cursor: pointer;
     background-color: rgb(195, 193, 193);
 }
-
-.text-based-header {
-    display: flex;
-    width: calc(100% - 20px);
-    height: fit-content;
-    flex-direction: row;
-    justify-content: space-between;
-    border-bottom: 1px solid rgba(89, 38, 16, 1);
-    padding: 12px 0px;
-    box-sizing: border-box;
-    margin: 0px 8px;
-}
-
-.question-header {
-    margin-top: 24px;
-    border-bottom: 1px solid black;
-    width: fit-content;
-}
-
-.result-section {
-    margin-left: 12px;
-    height: fit-content;
-    width: fit-content;
-}
-
-.text-results {
-    display: flex;
-    flex-direction: row;
-}
-
-.sidebar-title {
-    text-align: center;
-    background-color: rgb(236, 236, 236);
-    border-radius: 12px 12px 0px 0px;
-    padding: 12px 0px;
-    font-size: 20px;
-    margin-bottom: 0px;
-}
-
-.sidebar-options {
-    margin-left: 12px;
-    margin-top: 12px;
-    min-width: 200px;
-}
-
-.sidebar-options .nav-list {
-    min-height: 400px;
-    margin-top: 0px;
-}
-
-.text-scenario-results {
-    width: 100%;
-    display: grid;
-    grid-template-columns: 50% 50%;
-}
-
-.chart-scenario-results {
-    width: 100%;
-    display: grid;
-    grid-template-columns: 50% 50%;
-}
-
-.no-data {
-    margin-left: 12px;
-    margin-top: 12px;
-}
-
-.full-width-graph * {
-    width: 100% !important;
-}
-
-.full-width-graph .sa-question__title {
-    margin-bottom: 0px;
-    font-weight: 500;
-    font-family: var(--bs-body-font-family);
-    font-size: 16px;
-}
-
-.full-width-graph .sa-question .sa-visualizer__content {
-    border: 0.5px solid lightgrey;
-}
-
-.result-display {
-    width: calc(100% - 220px);
-}
-
-.viewGroup {
-    text-align: center;
-    align-items: center;
-}
-
-.count-column {
-    min-width: 100px;
-    width: 100px;
-}
-
-.answer-column {
-    width: 350px;
-    min-width: 350px;
-    max-width: 350px;
-}
-
-.graph-section .question-header {
-    margin-left: 28px;
-    margin-bottom: 0px;
-}
-
-.text-result-table {
-    width: fit-content;
-}
-
-.text-result-table thead {
-    position: static;
-}
-
-.by-participant td, .by-participant th {
-    min-width: 200px;
-    width: fit-content;
-    height: fit-content;
-}
-
-.table-container {
-    margin: 12px;
-    max-height: 700px;
-    overflow: auto;
-    border: 1px solid black;
-}
-
-.participant-text-results {
-    width: 100%;
-    display: flex;
-    flex-direction: column;
-}
-
-.participant-text-results button {
-    width: fit-content;
-    padding: 8px;
-    margin: 8px 0px 0px 12px;
-    border-radius: 4px;
-}
-
-.participant-text-results button:hover {
-    filter: brightness(90%);
-}
-
-.participant-text-results button:active {
-    filter: brightness(80%);
-}
